Changed color generation to chroma js based
Based on http://tools.medialab.sciences-po.fr/iwanthue/, uses the same JS to generate a palette for groups.
Can use the iwanthue site to experiment with different hcl limits to put on it for different kinds of palettes
Eventually it should be able to be more customizable, but I need to think about how to do that best.